backend/requirements-test.txt
# Testing framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1

# Test data generation
factory-boy==3.3.0
faker==20.1.0

# API testing
pytest-drf==0.0.1
pytest-html==4.1.1

# Performance testing
pytest-benchmark==4.0.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Security testing
bandit==1.7.5

# Database testing
pytest-mongodb==2.2.0 